Glow cells on the periodic table when moving.
Refactor in MagnetInteractions => Added a table attribute.
Refactor in MagnetInteractions => Added a Initialize(key) method to initialize the magnet with the key and its table parent.
Refactor in MagnetInteractions => Removed all call to parent table and replaced them to a reference to the table attribute.

Added glow game object reference that can be turned on and off in AtomCell. It is used used to highlight cells we can click on.

Added all functions in PeriodicTableMaze for turning highlighting on and off for the cells.